+++ src/index.ts
@@ -1,4 +1,5 @@
 export { Caller } from './Caller';
-export { extCall, setExtCallDefaults } from './ExtCall';
+export * from './ExtCall'; // All from it
 export type { obj } from './utils';
 export type { Handler, HandlerF } from './types';
+export { setFallbackLanguage, ErrorsMessagesDict } from './i18n/i18n';