+++ src/index.ts
@@ -1,5 +1,5 @@
 export { Caller } from './Caller';
 export * from './ExtCall'; // All from it
 export type { obj } from './utils';
-export type { Handler, HandlerF } from './types';
+export type { Handler, HandlerF, ExtError } from './types';
 export { setFallbackLanguage, createErrorsDict } from './i18n/i18n';